combine tests in one file
